Refactor rate limiting logic to support HTTP method-specific limits, enhance debug logging, and bump version to 5.4.0.

+++ src/main/java/dev/coph/simplerequest/util/Time.java
@@ -86,7 +86,7 @@ public class Time {
      *
      * @return the time value in seconds as an integer
      */
-    public int toSeconds() {
+    public long toSeconds() {
         return Math.toIntExact(milliseconds / 1000);
     }
